Refactored directory structure to be more clear
