+++ src/shared/resource/Resource.ts
